requirements.txt
pyrogram
pyromod==1.5
dnspython
aiohttp
TgCrypto
uvloop==0.19.0
pymongo==4.6.1
python-dotenv==1.0.0
apscheduler
pytz